requirements.txt
PyPDF2==3.0.1
pdf2image==1.16.3
pytesseract==0.3.10
numpy==1.24.3
Pillow==10.0.0
opencv-python==4.8.0.74
regex==2023.6.3
tqdm==4.65.0
Flask==2.3.3
Werkzeug==2.3.7 